requirements.txt
blinker==1.6.2
click==8.1.6
colorama==0.4.6
Flask==2.3.2
Flask-Cors==4.0.0
gunicorn==21.2.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
openpyxl==3.1.2
packaging==23.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Werkzeug==2.3.6
zipp==3.16.2
